requirements.txt
jupyter
pandas
numpy
matplotlib
seaborn
Pillow
scikit-learn
tqdm
gdown
torch
torchvision
open_clip_torch
pyarrow
